# docs/swagger.yaml
basePath: /api
definitions:
...
  APIError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  APIErrors:
    properties:
      code:
        type: integer
      errors:
        items:
          $ref: '#/definitions/APIError'
        type: array
    type: object
...
  Asset:
    properties:
      amount:
        type: string
      amount_usd:
        type: string
      currency_id:
        type: string
      identity:
        type: string
    type: object
...
  AuthRequest:
    properties:
      email:
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
      remember_me:
        type: boolean
    required:
    - email
    - password
    type: object
  AuthResponse:
    properties:
      token:
        type: string
    type: object
  Balance:
    properties:
      native_token:
        type: string
      native_token_usd:
        type: string
    type: object
...
  Blockchain:
    enum:
    - arbitrum
    - bitcoin
    - bitcoincash
    - bsc
    - dogecoin
    - ethereum
    - litecoin
    - polygon
    - tron
    type: string
    x-enum-varnames:
    - BlockchainArbitrum
    - BlockchainBitcoin
    - BlockchainBitcoinCash
    - BlockchainBinanceSmartChain
    - BlockchainDogecoin
    - BlockchainEthereum
    - BlockchainLitecoin
    - BlockchainPolygon
    - BlockchainTron
  BlockchainAdditionalData:
    properties:
      evm_data:
        $ref: '#/definitions/github.com/dv-net/dv-merchant_internal_service_wallet.EVMData'
      tron_data:
        $ref: '#/definitions/TronData'
    type: object
...
  CurrencyShort:
    properties:
      blockchain:
        $ref: '#/definitions/Blockchain'
      code:
        type: string
      id:
        type: string
      is_bitcoin_like:
        type: boolean
      is_evm_like:
        type: boolean
      name:
        type: string
      precision:
        type: integer
    type: object
...
  ExchangeSlug:
    enum:
    - htx
    - okx
    - binance
    - bitget
    - kucoin
    - bybit
    - gate
    type: string
    x-enum-varnames:
    - ExchangeSlugHtx
    - ExchangeSlugOkx
    - ExchangeSlugBinance
    - ExchangeSlugBitget
    - ExchangeSlugKucoin
    - ExchangeSlugBybit
    - ExchangeSlugGateio
...
  JSONResponse-AuthResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/AuthResponse'
      message:
        type: string
    type: object
...
  JSONResponse-ExternalProcessingWalletBalanceResponse:
    properties:
      code:
        type: integer
      data:
        items:
          $ref: '#/definitions/ProcessingWalletWithAssets'
        type: array
      message:
        type: string
    type: object
...
  JSONResponse-RegisterRootResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/RegisterRootResponse'
      message:
        type: string
    type: object
  JSONResponse-RegisterUserResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/RegisterUserResponse'
      message:
        type: string
    type: object
...
  JSONResponse-SearchByCriteriaResponse-any:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/SearchByCriteriaResponse-any'
      message:
        type: string
    type: object
...
  ProcessingWalletWithAssets:
    properties:
      additional_data:
        $ref: '#/definitions/BlockchainAdditionalData'
      address:
        type: string
      assets:
        items:
          $ref: '#/definitions/Asset'
        type: array
      balance:
        $ref: '#/definitions/Balance'
      blockchain:
        $ref: '#/definitions/Blockchain'
      currency:
        $ref: '#/definitions/CurrencyShort'
    type: object
...
  RegisterOwnerInfo:
    properties:
      owner_id:
        type: string
    type: object
  RegisterRequest:
    properties:
      cf-turnstile-response:
        type: string
      email:
        type: string
      language:
        maxLength: 2
        minLength: 2
        type: string
      location:
        type: string
      mnemonic:
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
    required:
    - email
    - language
    - location
    - password
    type: object
  RegisterRootResponse:
    properties:
      token:
        type: string
    type: object
  RegisterUserResponse:
    properties:
      token:
        type: string
      user_info:
        $ref: '#/definitions/github.com/dv-net/dv-merchant_internal_service_user.RegisterUserDTO'
    type: object
...
  SearchByCriteriaResponse-any:
    properties:
      data: {}
      search_type:
        type: string
    type: object
...
  TronData:
    properties:
      available_bandwidth_for_use:
        type: string
      available_energy_for_use:
        type: string
      stacked_bandwidth:
        type: string
      stacked_bandwidth_trx:
        type: string
      stacked_energy:
        type: string
      stacked_energy_trx:
        type: string
      stacked_trx:
        type: string
      total_bandwidth:
        type: string
      total_energy:
        type: string
      total_used_bandwidth:
        type: string
      total_used_energy:
        type: string
      tron_transfer_data:
        $ref: '#/definitions/TronTransferData'
    type: object
  TronTransferData:
    properties:
      max_transfers_native:
        type: string
      max_transfers_trc20:
        type: string
    type: object
...
  github.com/dv-net/dv-merchant_internal_models.RateSource:
    enum:
    - okx
    - htx
    - binance
    - bitget
    - coingate
    - dv-min
    - dv-max
    - dv-avg
    - kucoin
    - bybit
    - gate
    type: string
    x-enum-varnames:
    - RateSourceOKX
    - RateSourceHTX
    - RateSourceBinance
    - RateSourceBitGet
    - RateSourceCoingate
    - RateSourceDVMin
    - RateSourceDVMax
    - RateSourceDVAvg
    - RateSourceKucoin
    - RateSourceBybit
    - RateSourceGateio
  github.com/dv-net/dv-merchant_internal_models.User:
    properties:
      banned:
        $ref: '#/definitions/pgtype.Bool'
      created_at:
        type: string
      deleted_at:
        type: string
      dvnet_token:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      exchange_slug:
        $ref: '#/definitions/ExchangeSlug'
      id:
        type: string
      language:
        type: string
      location:
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
      processing_owner_id:
        type: string
      rate_scale:
        type: number
      rate_source:
        $ref: '#/definitions/github.com/dv-net/dv-merchant_internal_models.RateSource'
      remember_token:
        type: string
      updated_at:
        type: string
    required:
    - email
    - location
    - password
    type: object
...
  github.com/dv-net/dv-merchant_internal_service_user.RegisterUserDTO:
    properties:
      owner_info:
        $ref: '#/definitions/RegisterOwnerInfo'
      user:
        $ref: '#/definitions/github.com/dv-net/dv-merchant_internal_models.User'
    type: object
  github.com/dv-net/dv-merchant_internal_service_wallet.EVMData:
    properties:
      cost_per_erc20:
        description: Cost in Wei for single ERC20 transfer
        type: string
      cost_per_native:
        description: Cost in Wei for single native transfer
        type: string
      is_l2:
        description: Whether this is an L2 chain
        type: boolean
      l1_data_fee_estimate:
        description: Estimated L1 data fee for L2 chains
        type: string
      max_transfers_erc20:
        type: string
      max_transfers_native:
        type: string
      suggested_gas_price:
        type: string
    type: object
...
  pgtype.Bool:
    properties:
      bool:
        type: boolean
      valid:
        type: boolean
    type: object
...
info:
  contact:
    email: support@dv.net
    name: DV Support
  description: This is an API for DV Merchant
  title: DV Merchant
  version: "1.0"
paths:
...
  /v1/dv-admin/auth/login:
    post:
      consumes:
      - application/json
      description: Auth a user
      parameters:
      - description: Register account
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIErrors'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/APIErrors'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Auth user
      tags:
      - Auth
  /v1/dv-admin/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Register account
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-RegisterUserResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/APIErrors'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Register user
      tags:
      - Auth
  /v1/dv-admin/auth/register-root:
    post:
      consumes:
      - application/json
      description: Register root user - available only for first user
      parameters:
      - description: Register account
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-RegisterRootResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/APIErrors'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/APIErrors'
      summary: Register root user
      tags:
      - Auth
...
  /v1/dv-admin/search/{searchParam}:
    get:
      consumes:
      - application/json
      description: Search tx or wallet information in system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-SearchByCriteriaResponse-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrors'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - BearerAuth: []
      summary: Get wallet/tx data
      tags:
      - Search
...
  /v1/external/processing-wallet-balances:
    get:
      consumes:
      - application/json
      description: Get external processing wallet balances
      parameters:
      - description: Store API key
        in: query
        name: api_key
        type: string
      - collectionFormat: csv
        in: query
        items:
          enum:
          - arbitrum
          - bitcoin
          - bitcoincash
          - bsc
          - dogecoin
          - ethereum
          - litecoin
          - polygon
          - tron
          type: string
        name: blockchains
        type: array
        uniqueItems: true
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: currencies
        type: array
        uniqueItems: true
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/JSONResponse-ExternalProcessingWalletBalanceResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/APIErrors'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIErrors'
      security:
      - XApiKey: []
      summary: Get external processing wallet balances
      tags:
      - Withdrawal
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  XApiKey:
    description: Store API key
    in: header
    name: X-Api-Key
    type: apiKey
swagger: "2.0"